Feedback Part 5 - Move TabItemViewHolder and TabTrayAdapter to its own file
